♻️ refactor: 프로필 이미지 변경 기능 분리

+++ src/main/java/com/grepp/spring/app/controller/api/member/payload/MemberInfoResponse.java
@@ -10,7 +10,7 @@ public class MemberInfoResponse {
     private String id;
     private String email;
     private String name;
-    private Long profileImageNumber;
+    private Integer profileImageNumber;
     private String provider;
     private String role;
